backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
mysqlclient>=2.2
pymysql>=1.1
chromadb>=0.4
sentence-transformers>=2.2
groq>=0.4
selenium>=4.15
webdriver-manager>=4.0
undetected-chromedriver>=2.1
python-dotenv>=1.0
requests>=2.31
gunicorn>=21.0